requirements.txt
python-dotenv==1.0.1
Flask==3.0.3
requests-2.32.5
psycopg2-binary==2.9.9
boto3==1.34.120
botocore==1.34.120